Added encapsulation with VLAN hack
This commit implements encapsulation at the TC level using
bpf_skb_vlan_push(). This is basically a hack that allows us
to stack units of 4-bytes at a time. As a consequence, now all
our setup needs to use VLAN. Also, the re-encapsulated packet at
the end has a 2-byte leftover after the NSH, becasue 22 (size of
NSH + Outer Ethernet) is not a multiple of 4.