Improved test setup to multi-machine and switched to libvirt
Vagrantfile was severely changed, to reflect the actual test
environment we will use. Not it comprises of several VMs, instead
of just one.

We had to move from VirtualBox to libvirt because we needed to use
virtio driver with special flags at VMs to enable native XDP. We
only found documentation on how to it using QEMU, so we switched
over.