Fixed packet drops after XDP decap
For some reason, when we called bpf_xdp_adjust_head() to remove a
number of bytes not multiple by 8, the packet was always dropped
after BPF execution. Then we changed the number of extra bytes
after the NSH header from 2 to 6, this way our entire external
encapsulation is 32 bytes, which does cause issues with the
function anymore.