Add target to generate jited output code
This commit adds a new target `jited-out` to the Makefile which
is capable of outputing the jited BPF code + source annotations
to help debugging programs with perf, since we can't seem annotate
TC programs for now.

With these files, we can collect the annotated data using perf,
generate the jited output and compare both to know which parts of
the programs an running slowly.