Add helper code to interact with eBPF
This commit introduces cb_agent_helpers.*, which are supposed to be
compiled into a library to be used with the cb_agent_node, which is
a golang implementation of a agent to run inside each box receiving
configuration from the controller and performing any changes needed
to the host.

I chose to put implement all BPF-related interaction in C to make
use of libbpf, and just wrap it in some minimal library functions
that can easily be called from go.

Proper Makefile support to compile the agent_helper code into a lib
(libcbox) will be added on future commit (really soon). For now,
the library can be compiled by hand:

$ cd src/
$ gcc -I./utils/ -I./libbpf/src/root/usr/include/ -L./libbpf/src/ \
    -c cb_agent_helpers.c -lbpf -lelf
$ ar ruv libcbox.a cb_agent_helpers.o utils/load_helpers.o \
    utils/sys_helpers.o
$ cd control
$ ranlib libcbox.a
$ go build cb_agent_node.go