dataplane: Split Enc and Fwd into 2 BPF programs
Previously, Enc stage was a function called at the beginning of
Fwd stage, which made not possible to atomically replace only one
of the two programs to change functionality.

This commit finally splits these two logically independent stages
into two separate BPF programs.

For that, I *DID NOT* have to use tail calls. TC has a way to chain
classifiers, whose order can be specified using the 'prio' and
'handle' params. So when installing the programs, we set a 'prio'
value of 1 to Enc and 2 to Fwd. TC will execute then in ascending
order. Fallthrough to the next program que be specified inside BPF
by returning TC_ACT_UNSPEC, which is done at the end of Enc stage
to guarantee that TC will execute Fwd stage right after it.