control: split configuration of rx and tx stages
In order to properly use veth pairs in the container scenario, we
need to be able to *not* load some ingress/egress stages when using
veth pairs. So I had to split the loading of all stages in two, to
handle rx (ingress) and tx (egress) separately.

Now the agent also allows using a different ifaces for rx and tx,
to support SFs that might redirect packets to a different interface.