docker: Error out when misusing direct links
The use of direct links to connect one SF to many other SFs that
go after it on any chains can lead to a 1-to-N or N-to-1 relationship
which the current architecture cannot handle. We would have to allow
receiving packets and sending packets to many interfaces, and most
likely running our SF applications for many RX interfaces as well,
which would increase complexity.

For now, when using direct links, the container *should not* be
shared between chains. Instead, it should be dedicated to a single
chain.

This only applies when using direct links. If all SFs in a config
file are 'remote', then no DLs will be used and all is fine, i.e.
the SF can be shared between as many chains as you wish.