Allow disabling stages on direct links
When using direct links (veth pairs) in the container scenario, we
don't need the BPF stages to do encap/decap/forwarding, since the
ingress/egress interface will directly deliver the packet to the
next function.

This commit enables passing '-' as value to --ingress and --egress
params to cb_start / cb_agent so the agent does not install the
stages on the corresponding interface.

This shows a performance improvement, as less processing on the
packet is needed when going through direct links.