Generate proxy rules + fwd rules only when needed
This commit gives the manager the ability to create artificial
proxy rules in the two cases where they are needed (see comment in
src/control/cbox/manager.go).

In addition, it makes the manager only generate forwarding rules when
we do not have a direct link to the next chain.