libvirt: Fix password-less SSH connection
Even though the user's key was being correctly installed in the images, it was
unable to SSH into the VMs. This was being caused because there were no SSH host
keys installed in the VMs. They got removed by the 'ssh-keys' operation from
virt-sysprep.

So make sure key generation the first time the VM boots.